+++ Network/Telegram/API/Bot/Object/From.hs
@@ -3,7 +3,15 @@ module Network.Telegram.API.Bot.Object.From
 
 import "aeson" Data.Aeson (FromJSON (parseJSON), Object, withObject, (.:), (.:?))
 import "aeson" Data.Aeson.Types (Parser)
-import "base" Data.Bool (bool)
+import "base" Control.Applicative (Applicative ((<*>)))
+import "base" Control.Monad (Monad ((>>=)))
+import "base" Data.Bool (Bool (False), bool)
+import "base" Data.Eq (Eq ((==)))
+import "base" Data.Int (Int)
+import "base" Data.Function (($))
+import "base" Data.Functor ((<$>))
+import "base" Data.Maybe (Maybe)
+import "base" Text.Show (Show)
 import "lens" Control.Lens (Lens')
 import "text" Data.Text (Text)