+++ CHANGELOG.md
@@ -23,3 +23,4 @@
 * Define `Member` datatype for chat users
 * No implicit `Prelude` anymore (NoImplicitPrelude)
 * Move HTTP request function into `Internal` module
+* Define `Object` type family